third_party/WebKit/LayoutTests/fast/files/blob-close-expected.txt

Test the Blob.close() method, basic functionality.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS blob instanceof window.Blob is true
PASS blob.size is 5
PASS blob.isClosed is false
PASS blob.close(); blob.size is 0
PASS blob.isClosed is true
PASS blob.type is "text/plain"
PASS sliced1.size is 3
PASS blob.isClosed is false
PASS blob.close(); blob.size is 0
PASS blob.isClosed is true
PASS blob.type is "text/plain"
PASS sliced1.size is 3
PASS sliced2.size is 1
PASS blob.slice(2) threw exception InvalidStateError: Failed to execute 'slice' on 'Blob': Blob has been closed..
PASS blob.close() threw exception InvalidStateError: Failed to execute 'close' on 'Blob': Blob has been closed..
PASS successfullyParsed is true

TEST COMPLETE
